feat(OUT-1785): implement task update in new realtime handler
